refactor: rename/reorganize files & remove action/method from form
